pyproject.toml
[project]
name = "nextbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.2",
    "django>=5.2.8",
    "django-ninja>=1.4.5",
    "editorconfig-checker>=3.4.1",
    "environs[django]>=14.5.0",
    "polars>=1.35.2",
    "pydantic-settings>=2.11.0",
    "python-telegram-bot>=22.5",
    "ruff>=0.14.4",
    "xlsxwriter>=3.2.9",
]
